+++ seqkit/cmd/faidx.go
@@ -57,7 +57,7 @@ This command is similar with "samtools faidx" but has some extra features:
 		ignoreCase := getFlagBool(cmd, "ignore-case")
 		useRegexp := getFlagBool(cmd, "use-regexp")
 
-		files := getFileList(args)
+		files := getFileList(args, false)
 
 		file := files[0]