+++ seqkit/cmd/fish.go
@@ -114,7 +114,13 @@ func parseRanges(rf string) Ranges {
 var fishCmd = &cobra.Command{
 	Use:   "fish",
 	Short: "look for short sequences in larger sequences using local alignment",
-	Long:  "look for short sequences in larger sequences using local alignment",
+	Long:  `look for short sequences in larger sequences using local alignment
+
+Attention:
+  1. output coordinates are BED-like 0-based, left-close and right-open.
+  2. alignment information are printed to STDERR.
+
+`,
 
 	Run: func(cmd *cobra.Command, args []string) {
 		config := getConfigs(cmd)
@@ -158,7 +164,7 @@ var fishCmd = &cobra.Command{
 			seq.ValidateSeq = true
 		}
 
-		files := getFileList(args)
+		files := getFileList(args, true)
 		var alns []*AlignedSeq
 		if len(files) == 0 {
 			files = []string{"-"}