+++ seqkit/cmd/sana.go
@@ -60,7 +60,7 @@ var sanaCmd = &cobra.Command{
 		defer outfh.Flush()
 		defer outfh.Close()
 
-		files := getFileList(args)
+		files := getFileList(args, true)
 
 		for _, file := range files {
 			rawSeqChan := NewRawSeqStreamFromFile(file, 1000, qBase, "fasta")