+++ seqkit/cmd/shuffle.go
@@ -63,7 +63,7 @@ Secondly, seqkit shuffles sequence IDs and extract sequences by FASTA index.
 		fai.MapWholeFile = false
 		runtime.GOMAXPROCS(config.Threads)
 
-		files := getFileList(args)
+		files := getFileList(args, true)
 		seed := getFlagInt64(cmd, "rand-seed")
 		twoPass := getFlagBool(cmd, "two-pass")
 		keepTemp := getFlagBool(cmd, "keep-temp")