+++ seqkit/cmd/split2.go
@@ -66,7 +66,7 @@ according to the input files.
 		fai.MapWholeFile = false
 		runtime.GOMAXPROCS(config.Threads)
 
-		files := getFileList(args)
+		files := getFileList(args, true)
 
 		if len(files) > 1 {
 			checkError(fmt.Errorf("no more than one file should be given"))