+++ CHANGES.md
@@ -1,3 +1,10 @@
+- [SeqKit v0.11.1](https://github.com/shenwei356/seqkit/releases/tag/v0.11.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.11.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.11.1)
+    - `seqkit`: fix checking input file existence.
+    - `seqkit seq`: do not force switch on `-g` when using `-m/-M`.
+    - `seqkit translate`: supporting multiple frames. [#96](https://github.com/shenwei356/seqkit/issues/96)
+    - `seqkit grep/locate`: add detection and warning for space existing in search pattern/sequence.
+    - `seqkit version`: makeing checking update optional.
 - [SeqKit v0.11.0](https://github.com/shenwei356/seqkit/releases/tag/v0.11.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.11.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.11.0)
     - `seqkit`: fix hanging when reading from truncated gzip file.